+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 [...]
 
+- **[UPDATE]** `SlideSection` behavior changed for small screens
+
 # v41.4.0 (19/10/2020)
 
 - **[NEW]** Add support for onClick handler for `Caption` link.